fraction refactor to fix bug
