+++ lib/expression/step-solver/LikeTermCollector.js
@@ -182,7 +182,7 @@ function getTermsForCollectingAddition(node) {
       const termName = getTermName(child, '+');
       terms = Util.appendToArrayInObject(terms, termName, child);
     }
-    else if (Fraction.isIntegerFraction(child)) {
+    else if (NodeType.isIntegerFraction(child)) {
       terms = Util.appendToArrayInObject(terms, CONSTANT_FRACTION, child);
     }
     else if (NodeType.isConstant(child)) {
@@ -233,7 +233,7 @@ function getTermsForCollectingMultiplication(node) {
     if (PolynomialTermNode.isPolynomialTerm(child)) {
       terms = addToTermsforPolynomialMultiplication(terms, child);
     }
-    else if (Fraction.isIntegerFraction(child)) {
+    else if (NodeType.isIntegerFraction(child)) {
       terms = Util.appendToArrayInObject(terms, CONSTANT, child);
     }
     else if (NodeType.isConstant(child)) {
