+++ lib/expression/step-solver/Negative.js
@@ -33,6 +33,9 @@ Negative.isNegative = function(node) {
 }
 
 Negative.negate = function(node) {
+  if (!node) {
+    return NodeCreator.constant(-1);
+  }
   if (NodeType.isUnaryMinus(node)) {
     return node.args[0];
   }
