+++ lib/expression/step-solver/NodeType.js
@@ -62,4 +62,13 @@ NodeType.isConstantOrConstantFraction = function(node, allowUnaryMinus=false) {
   }
 }
 
+NodeType.isIntegerFraction = function(node) {
+  if (!NodeType.isConstantFraction(node)) {
+    return false;
+  }
+  return (Number.isInteger(parseFloat(node.args[0].value)) &&
+          Number.isInteger(parseFloat(node.args[1].value)));
+}
+
+
 module.exports = NodeType;
