+++ lib/expression/step-solver/PolynomialTermOperations.js
@@ -2,7 +2,7 @@
 // Operations on polynomial term nodes
 
 const math = require('../../../index');
-const Fraction = require('./Fraction');
+const ConstantFraction = require('./ConstantFraction');
 const MathChangeTypes = require('./MathChangeTypes');
 const NodeCreator = require('./NodeCreator');
 const NodeStatus = require('./NodeStatus');
@@ -113,7 +113,7 @@ PolynomialTermOperations.simplifyPolynomialFraction = function(node) {
   }
 
   const coefficientFraction = polyNode.getCoeffNode(); // a division node
-  const newCoeffStatus = Fraction.simplifyFraction(coefficientFraction);
+  const newCoeffStatus = ConstantFraction.simplifyFraction(coefficientFraction);
   if (!newCoeffStatus.hasChanged) {
     return new NodeStatus(node);
   }
