+++ lib/expression/step-solver/cancelLikeTerms.js
@@ -3,6 +3,7 @@
 const math = require('../../../index');
 
 const MathChangeTypes = require('./MathChangeTypes');
+const Negative = require('./Negative');
 const NodeCreator = require('./NodeCreator');
 const NodeStatus = require('./NodeStatus');
 const NodeType = require('./NodeType');
@@ -172,7 +173,7 @@ function cancelTerms(numerator, denominator) {
   if (NodeType.isUnaryMinus(numerator)) {
     const cancelStatus = cancelTerms(numerator.args[0], denominator);
     // if numerator and denominator are both unary minus, it cancels out
-    if (NodeType.isUnaryMinus(denominator)) {
+    if (NodeType.isUnaryMinus(denominator) || !cancelStatus.numerator) {
       numerator = cancelStatus.numerator;
     }
     else {
@@ -182,11 +183,17 @@ function cancelTerms(numerator, denominator) {
     return new CancelOutStatus(numerator, denominator, cancelStatus.hasChanged);
   }
   if (NodeType.isUnaryMinus(denominator)) {
-    cancelStatus = cancelTerms(numerator, denominator.args[0]);
+    const cancelStatus = cancelTerms(numerator, denominator.args[0]);
+    numerator = cancelStatus.numerator;
     // note: if the nuemrator was unary minus, it'd already be caught with the
     // above code block
-    denominator.args[0] = cancelStatus.denominator;
-    numerator = cancelTerms.numerator;
+    if (cancelStatus.denominator) {
+      denominator.args[0] = cancelStatus.denominator;
+    }
+    else {
+      denominator = cancelStatus.denominator;
+      numerator = Negative.negate(numerator);
+    }
     return new CancelOutStatus(numerator, denominator, cancelStatus.hasChanged);
   }
