+++ lib/expression/step-solver/prettyPrint.js
@@ -37,7 +37,7 @@ function prettyPrintDFS(node, latex) {
       return `${coeffStr} ${nonCoeffStr}`;
     }
   }
-  if (Fraction.isIntegerFraction(node)) {
+  if (NodeType.isIntegerFraction(node)) {
     if (latex) {
       // fractions in latex are different than in asciimath
       return `\\frac{${node.args[0]}}{${node.args[1]}}`;
@@ -104,6 +104,7 @@ function prettyPrintDFS(node, latex) {
   }
   else if (NodeType.isUnaryMinus(node)) {
     if (NodeType.isOperator(node.args[0]) &&
+        node.args[0].op !== '/' &&
         !PolynomialTermNode.isPolynomialTerm(node)) {
       return `-(${prettyPrintDFS(node.args[0], latex)})`;
     }
