+++ lib/expression/step-solver/removeUnnecessaryParens.js
@@ -116,7 +116,7 @@ function removeUnnecessaryParensInParenthesisNode(node) {
   // If the content is just one symbol or constant, the parens are not
   // needed.
   else if (NodeType.isConstant(node.content, true) ||
-           Fraction.isIntegerFraction(node.content) ||
+           NodeType.isIntegerFraction(node.content) ||
            NodeType.isSymbol(node.content)) {
     node = node.content;
     hasChanged = true;
