+++ lib/expression/step-solver/simplifyOperations.js
@@ -3,6 +3,7 @@
 const math = require('../../../index');
 
 const cancelLikeTerms = require('./cancelLikeTerms');
+const ConstantFraction = require('./ConstantFraction');
 const Fraction = require('./Fraction');
 const MathChangeTypes = require('./MathChangeTypes');
 const NodeCreator = require('./NodeCreator');
@@ -85,7 +86,7 @@ function simplifyOperationsBeforeRecursion(node) {
     removeDivisionByOne,
     // multiplication by 0 yields 0
     reduceMultiplicationByZero,
-    Fraction.addConstantFractions,
+    ConstantFraction.addConstantFractions,
   ];
 
   for (let i = 0; i < simplificationFunctions.length; i++) {
@@ -111,8 +112,8 @@ function simplifyOperationsAfterRecursion(node) {
     removeMultiplicationByNegativeOne,
 
     // FRACTION simplifications (for fractions with only constant nodes)
-    Fraction.multiplyConstantsAndFractions,
-    Fraction.addConstantAndFraction,
+    ConstantFraction.multiplyConstantsAndFractions,
+    ConstantFraction.addConstantAndFraction,
     Fraction.simplifyFraction,
     Fraction.multiplyByInverse,
 
@@ -186,7 +187,7 @@ function performArithmetic(node) {
 
   // Only resolve division of integers if we get an integer result.
   // Note that a fraction of decimals will be divided out.
-  if (Fraction.isIntegerFraction(node)) {
+  if (NodeType.isIntegerFraction(node)) {
     const numeratorValue = parseInt(node.args[0]);
     const denominatorValue = parseInt(node.args[1]);
     if (numeratorValue % denominatorValue === 0) {
